+++ scala/src/main/scala/euler/til10/Euler9.scala
@@ -9,7 +9,7 @@ object Euler9 extends EulerProblem {
     val products = for {
       a <- 2 to 999
       b <- a + 1 to 999
-      c = sqrt(a * a + b * b)
+      c = sqrt(a.toDouble * a + b * b)
       if a + b + c == 1000
     } yield (a * b * c).toInt
